Separated tests
Tests that check the correctness of an implementation have
been separated from tests that check that an algorithm is
non-blocking.

This way, tests that need I/O are much lighter (100 cases)
because no more are needed to test that the algorithm works.

On the other hand, tests to search blocking code can keep a
high number of elements without impacting times.