Implemented all methods in WriteBufferManager.
Abandoned the idea of having a secondary buffer to cope with write
attempts during synchronization, in order not to miss a possible
match between those elements and those in the other side's write
buffer.

Need to think a solution for this. Probably, the idea of an
unpromotable secondary write buffer that serves for holding elements
to be written while the sync is on would be good. There would be a
post sync state in which both secondaries are matched against the
elements in the opposite write buffer. After that, write buffers
are cleared and populated with the unmatched elements coming from
their secondaries, which get cleared too.

This would probably not be necessary because in the sync phases, no
elements from the read buffers are being processed.